requirements.txt
openpyxl==3.1.5
pandas==2.2.3
Pillow==10.4.0
plotly==5.9.0
PuLP==2.8.0
PyYAML==6.0.2
Requests==2.32.3
streamlit==1.33.0